Fix link definition parsing being too greedy
Currently, the following snippet:

```
[foo][bar]: click me

[bar]: https://example.com
```

will be parsed as

```
Start(LinkDefinition { label: "foo][bar" }, {})
Str("click me")
End(LinkDefinition { label: "foo][bar" })
```

leading to somewhat surprising results. This happens not only when that
construct is the start of a paragraph, but also when it is the start of
a list item:

```
* [foo][bar]: click me
* [foo][bar] - click me, too

[bar]: https://example.com
```

First link does not render, the second one will. This behavior is rooted
in a somewhat too greedy approach to identifying the closing `]:` when a
link definition is possible.

This commit fixes that by asserting that `]:` is found at the _first_
occurence of `]` when looking for a LinkDefinition.

Signed-off-by: Conrad Hoffmann <ch@bitfehler.net>