Handle repeated attribute specifiers in pre-pass
While the regular parser handles repeated attribute specifiers just
fine, pre-pass does not - it only keeps track of the last attribute
specifier it encountered, leading to for example this snippet:

```
[foo][bar]: click me

{title="Example link"}
{rel="external"}
[bar]: https://example.com
```

being rendered as

```
<p><a href="https://example.com" rel="external">foo</a>: click me</p>
```

(note the missing title).

This commit fixes that by having the pre-pass keep track of a list of
attribute specifiers instead of just a single one.