Fix div being closed by raw block content
Currently a div gets closed if it contains a raw block that in turn
contains what could be the div's closing fence. This commit fixes that
by keeping track of open code blocks when looking for the closing fence.

It also removes the existing test for this case from the `skip` file.

fixes #6