Fix lexer eating ! before plain span
Currently, `foo![bar]{.cls}` will swallow the exclamation mark, because
it is first lexed as opener of an image link, then later the converted
to a plain span because it has attributes but no URL.

The reference implementation handles this correctly [1].

To mitigate, do not lex `![` as special symbol. Instead, check during
span-to-link conversion if the opener was preceeded by a `!` and modify
events accordingly if so.

[1] https://djot.net/playground/?text=%21%5Bspan%5D%7B.bar%7D&mode=html&sourcepos=false